Split FlexibleConfig into interface and implementation
